+++ src/com/kaylerrenslow/armaDialogCreator/arma/stringtable/StringTableKey.java
@@ -203,7 +203,6 @@ public interface StringTableKey extends Macro<SVString> {
 
 	@NotNull
 	default String getHeaderMacroId() {
-		System.out.println("StringTableKey.getHeaderMacroId getHeaderMacroId(this)=" + getHeaderMacroId(this));
 		return getHeaderMacroId(this);
 	}